extract LCP_context token to separate file. And use inject() function in directive
