+++ projects/storefrontlib/cms-structure/lcp-context/lcp-context.directive.ts
@@ -1,12 +1,14 @@
 import {
   Directive,
   inject,
+  OnInit,
   TemplateRef,
   ViewContainerRef,
 } from '@angular/core';
 import { distinctUntilChanged, map, Observable, shareReplay } from 'rxjs';
 import { ImageFetchPriority } from '../../shared/components/media/media.model';
-import { LCP_CONTEXT, LcpPresence } from './lcp-context.model';
+import { LcpPresence } from './lcp-context.model';
+import { LCP_CONTEXT } from './lcp-context.token';
 import { LcpPresenceMappingService } from './lcp-presence-mapping.service';
 
 interface LcpContextDirectiveTemplateContext {
@@ -20,24 +22,25 @@ interface LcpContextDirectiveTemplateContext {
   selector: '[cxLcpContext]',
   standalone: false,
 })
-export class LcpContextDirective {
+export class LcpContextDirective implements OnInit {
+  protected readonly lcpContext = inject(LCP_CONTEXT);
   protected readonly lcpPresenceMappingService = inject(
     LcpPresenceMappingService
   );
-  readonly lcpContext = inject(LCP_CONTEXT);
-  readonly fetchPriority$ = this.lcpContext.lcpPresence$.pipe(
-    map(
-      (lcpElementInfo) =>
-        this.lcpPresenceMappingService.getFetchPriority(lcpElementInfo) ?? null
+  protected readonly fetchPriority$ = this.lcpContext.lcpPresence$.pipe(
+    map((lcpElementInfo) =>
+      this.lcpPresenceMappingService.getFetchPriority(lcpElementInfo)
     ),
     distinctUntilChanged(),
     shareReplay({ bufferSize: 1, refCount: true })
   );
 
-  constructor(
-    private templateRef: TemplateRef<LcpContextDirectiveTemplateContext>,
-    private viewContainer: ViewContainerRef
-  ) {
+  protected templateRef = inject(
+    TemplateRef<LcpContextDirectiveTemplateContext>
+  );
+  protected viewContainer = inject(ViewContainerRef);
+
+  ngOnInit(): void {
     this.viewContainer.createEmbeddedView(this.templateRef, {
       $implicit: {
         lcpPresence$: this.lcpContext.lcpPresence$,
